+++ src/index.ts
@@ -1,10 +1,10 @@
-export * from './bold';
-export * from './code';
-export * from './codeblock';
-export * from './compose';
-export * from './italic';
-export * from './spoiler';
-export * from './strikethrough';
-export * from './underline';
-export * from './quote';
-export * from './constructTemplate';
+export * from './lib/bold';
+export * from './lib/code';
+export * from './lib/codeblock';
+export * from './lib/italic';
+export * from './lib/spoiler';
+export * from './lib/strikethrough';
+export * from './lib/underline';
+export * from './lib/quote';
+export * from './util/compose';
+export * from './util/constructTemplate';
